add presets: DRUIDS (Light Theme) and DRUIDS (Dark Theme)
These are based on https://druids.datadoghq.com/foundations/color

Because DRUIDS (Dark Theme) starts with the darkest colors at 100, I thought
about re-ordering the tones so that they are 700..100 instead, but I noticed
that Radix UI (Dark Theme) already exists as a preset with the tones ordered
from darkest to lightest. This might preserve the color designer's intent,
but it also makes it harder to compare visually against other presets.
If "dark theme" presets really should be added as-is with darkest tones first,
it might be useful to later have a "flip horizontally" button to make
comparisons between color presets easier.